GrayCoding

Documentation for GrayCoding.

GrayCoding.GrayMatrix

GrayCoding.bin2dec

GrayCoding.dec2bin

GrayCoding.dnamatrix

GrayCoding.gen_gray

GrayCoding.matrixplot

GrayCoding.pam_encode

GrayCoding.GrayMatrix — Method

Generate Encoding and Decoding matrices for Gray Codes of alphabet.

GrayCoding.bin2dec — Method

Binary to decimal number conversion. Input can be

binary strings,

binary digits or

a vector of binary string or digits

julia> bin2dec([011,"0111",0111])

GrayCoding.dec2bin — Method

Decimal to binary conversion

GrayCoding.dnamatrix — Method

Plot the DNA codon matrix

julia> dnamatrix()

GrayCoding.gen_gray — Method

Generate Gray vectors

GrayCoding.matrixplot — Method

Plots a matrix into a 2D with labels. Optional arguments including colors `$julia-repl julia> using Random;A= rand(0:9,10,10);matrixplot(A)$

GrayCoding.pam_encode — Method

Pulse amplitude modulation (PAM) mapping. This is a type of digital modulation mapping used in Communication systems.