GrayCoding

Documentation for GrayCoding.

Introduction

Welcome to the documentation for GrayCoding!

What is GrayCocoding.jl?

GrayCoding is a formal Linear Algebraic framework for $q$-ry Gray Code. Encoding and Decooding of Gray codes can be treated as a special case of algebraic block coding.

Encoding: $\textbf{g}=G \textbf{b}$

Decoding: $\textbf{b}=B \textbf{g}$

Tip

This is still under active devlopment.

Resources for getting started

There are few ways to get started with GrayCoding:

Read TBD.

GrayCoding.G

GrayCoding.Gn

GrayCoding.GrayMatrix

GrayCoding.bin2dec

GrayCoding.dec2bin

GrayCoding.dnamatrix

GrayCoding.gen_gray

GrayCoding.gray

GrayCoding.gray_recursion

GrayCoding.level2unitary

GrayCoding.matrixplot

GrayCoding.pam_encode

GrayCoding.reflect_code

GrayCoding.G — Method

Find the Givens embedding ${}^{i}G_{j,k}(A)$

GrayCoding.Gn — Method

Decomposition of aribtrary unitary matrix $U(n) \in S(2^n)$, as a cascade of two level Givens matrices.

Using the following property,

\[\prod_{j=0}^{n-1}{ {}^{1}G_{n-j,n-j-1} U(n)} = \begin{pmatrix} 1 & 0 \\ 0 & U(n-1)\end{pmatrix}\]

\[\prod_{i=1}^{n-1} \prod_{j=0}^{n-1-i}{ {}^{i}G_{n-j,n-j-1} U(n)} = \begin{pmatrix} 1 & \ldots & 0 \\ \vdots & \ddots & \vdots \\ 0 & \ldots & \det()\end{pmatrix}\]

Parameters

U – Input. Unitary matrix of size $2^n$

V – Output unitary matrix in two level form [1 0;0 U'] form where U' is a unitary matrix of size $2^{n-1}$.

Gm – The $(n-2)(n-1)$ sequence of Given matrices (in augmented form) $[{}^{1}G_{n}\lvert{}^{1}G_{n-1}\lvert\ldots\lvert{}^{1}G_{2}\lvert {}^{2}G_{n}\lvert{}^{2}G_{n-1}\lvert\ldots\lvert{}^{2}G_{3}\lvert\ldots\lvert{}^{n-2}G_{n}\lvert{}^{n-2}G_{n-1}\lvert{}^{n-1}G_{n}]$

Examples

julia> Gn(U)

GrayCoding.GrayMatrix — Function

Generate Encoding and Decoding matrices for Gray Codes of alphabet.

julia> G,B,g,b=GrayMatrix(4, 2);
julia> G
    4×4 Matrix{Int64}:
    1  0  0  0
    1  1  0  0
    1  1  1  0
    1  1  1  1
    julia> B
    4×4 Matrix{Int64}:
    1  0  0  0
    1  1  0  0
    0  1  1  0
    0  0  1  1
    julia> g 
    4×16 Matrix{Int64}:
    0  0  0  0  0  0  0  0  1  1  1  1  1  1  1  1
    0  0  0  0  1  1  1  1  0  0  0  0  1  1  1  1
    0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1
    0  1  0  1  0  1  0  1  0  1  0  1  0  1  0  1
    julia> b 
    4×16 Matrix{Int64}:
    0  0  0  0  0  0  0  0  1  1  1  1  1  1  1  1
    0  0  0  0  1  1  1  1  1  1  1  1  0  0  0  0
    0  0  1  1  1  1  0  0  0  0  1  1  1  1  0  0
    0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0

GrayCoding.bin2dec — Method

Binary to decimal number conversion. Input can be

binary strings,

binary digits or

a vector of binary string or digits

julia> bin2dec([011,"0111",0111])

GrayCoding.dec2bin — Method

Decimal to binary conversion

GrayCoding.dnamatrix — Method

Plot the DNA codon matrix

julia> dnamatrix()

GrayCoding.gen_gray — Method

Generate Gray vectors

GrayCoding.gray — Method

Recursive construction of binary Gray code digits.

Gray code $g[n]$ can be recursively constructed as follows. Start with $g[1] = (0,1) = (g_{1},g_{2})$

\[g[n+1] = 0g_{1},...,0g_{N−1},0g_{N},1g_{N},1g_{N−1},...,1g_{1}.\]

Examples

julia> gray(3)
3×8 Matrix{Int64}:
 0  0  0  0  1  1  1  1
 0  0  1  1  0  1  1  0
 0  1  1  0  1  1  0  0

GrayCoding.gray_recursion — Method

Recursive function to illustrate the reflection+shift property of Gray mapping.

Arguents

n - The iteration number n ≥ 0

C - The decimal sequence of the gray mapped bits

R - The reflected sequence (without shifting)

julia> C,R = gray_recursion(4)

GrayCoding.level2unitary — Method

Given unitary matrux $U(n) \in S(2^n)$, it uses a a repeated Givens rotations to get a to level matrix as follows.

\[\prod_{j=1}^{n-2}{ {}^{1}G_{n-j,n-j-1} U(n)} = \begin{pmatrix} 1 & 0 \\ 0 & U(n-1)\end{pmatrix}\]

Parameters

U – Input. Unitary matrix of size $2^n$

V – Output unitary matrix in two level form [1 0;0 U'] form where U' is a unitary matrix of size $2^{n-1}$.

GG – The $n-2$ sequence of Given matrices (in augmented form) $[{}^{1}G_{n}\lvert{}^{1}G_{n-1}\lvert\ldots\lvert{}^{1}G_{2}]$

Examples

julia> level2unitary(U)

GrayCoding.matrixplot — Method

Plots a matrix into a 2D with labels. Optional arguments including colors

julia> using Random;
julia> A= rand(0:9,10,10);
julia> matrixplot(A)

GrayCoding.pam_encode — Method

Pulse amplitude modulation (PAM) mapping. This is a type of digital modulation mapping used in Communication systems.

GrayCoding.reflect_code — Method

Reflected code.

julia>reflect_code(3)
[0,1,3,2,2,3,1,0]