GrayCoding

Documentation for GrayCoding.

Introduction

Welcome to the documentation for GrayCoding!

What is GrayCocoding.jl?

GrayCoding is a formal Linear Algebraic framework for $q$-ry Gray Code. Encoding and Decooding of Gray codes can be treated as a special case of algebraic block coding.

Encoding: $g=G*b$

Decoding: $b=B*g$

Tip

This is still under active devlopment.

Resources for getting started

There are few ways to get started with GrayCoding:

Read TBD.

GrayCoding.GrayMatrix

GrayCoding.bin2dec

GrayCoding.dec2bin

GrayCoding.dnamatrix

GrayCoding.gen_gray

GrayCoding.matrixplot

GrayCoding.pam_encode

GrayCoding.GrayMatrix — Function

Generate Encoding and Decoding matrices for Gray Codes of alphabet.

julia> G,B,g,b=GrayMatrix(4, 2);
julia> G
    4×4 Matrix{Int64}:
    1  0  0  0
    1  1  0  0
    1  1  1  0
    1  1  1  1
    julia> B
    4×4 Matrix{Int64}:
    1  0  0  0
    1  1  0  0
    0  1  1  0
    0  0  1  1
    julia> g 
    4×16 Matrix{Int64}:
    0  0  0  0  0  0  0  0  1  1  1  1  1  1  1  1
    0  0  0  0  1  1  1  1  0  0  0  0  1  1  1  1
    0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1
    0  1  0  1  0  1  0  1  0  1  0  1  0  1  0  1
    julia> b 
    4×16 Matrix{Int64}:
    0  0  0  0  0  0  0  0  1  1  1  1  1  1  1  1
    0  0  0  0  1  1  1  1  1  1  1  1  0  0  0  0
    0  0  1  1  1  1  0  0  0  0  1  1  1  1  0  0
    0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0

GrayCoding.bin2dec — Method

Binary to decimal number conversion. Input can be

binary strings,

binary digits or

a vector of binary string or digits

julia> bin2dec([011,"0111",0111])

GrayCoding.dec2bin — Method

Decimal to binary conversion

GrayCoding.dnamatrix — Method

Plot the DNA codon matrix

julia> dnamatrix()

GrayCoding.gen_gray — Method

Generate Gray vectors

GrayCoding.matrixplot — Method

Plots a matrix into a 2D with labels. Optional arguments including colors

julia> using Random;
julia> A= rand(0:9,10,10);
julia> matrixplot(A)

GrayCoding.pam_encode — Method

Pulse amplitude modulation (PAM) mapping. This is a type of digital modulation mapping used in Communication systems.